requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==4.2.3
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.6
pytest==8.3.3
sqlparse==0.5.1
